refactor(api): refactor API endpoints for user registration and login
