--- backend/api/serializer.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from user.models import User
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta(object):
-        model = User
-        fields = '__all__'
