+++ backend/config/urls.py
@@ -19,5 +19,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('user/', include('api.urls')),
+    path('api/v1/', include('api.urls')),
 ]
