+++ backend/user/serializer.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from .models import User
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta(object):
+        model = User
+        fields = '__all__'
