+++ bruno/search.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: {{base_url}}:{{port}}/user/search/
+  url: {{base_url}}:{{port}}/api/v1/search/
   body: json
   auth: none
 }
