+++ bruno/swagger.bru
@@ -1,0 +1,11 @@
+meta {
+  name: swagger
+  type: http
+  seq: 5
+}
+
+get {
+  url: {{base_url}}:{{port}}/api/v1/swagger
+  body: none
+  auth: none
+}
